requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
brotli==1.2.0
cbor==1.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
datasets==4.6.1
dill==0.4.0
distro==1.9.0
fastapi==0.135.1
ffmpy==1.0.0
filelock==3.25.0
FlagEmbedding==1.3.5
frozenlist==1.8.0
fsspec==2026.2.0
streamlit==1.42.0

groovy==0.1.2
groq==1.0.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ijson==3.5.0
inscriptis==2.7.1
ir_datasets==0.5.11
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numpy==2.4.2
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pdfminer.six==20251230
pdfplumber==0.11.9
peft==0.18.1
pillow==12.1.1
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.27.1
pypdfium2==5.5.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
sentence-transformers==5.2.3
sentencepiece==0.2.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.20.3
tomlkit==0.13.3
torch==2.10.0
tqdm==4.67.3
transformers==4.46.3
trec-car-tools==2.6
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unlzw3==0.2.3
urllib3==2.6.3
uvicorn==0.41.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
xxhash==3.6.0
yarl==1.23.0
zlib-state==0.1.10
